Fix alembic migration
