--- backend/alembic/versions/678f1b5ea40c_migration.py
@@ -1,36 +1,0 @@
-"""migration
-
-Revision ID: 678f1b5ea40c
-Revises: 4ae4ff0f990a
-Create Date: 2025-02-27 12:42:19.756651
-
-"""
-
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = "678f1b5ea40c"
-down_revision: Union[str, None] = "4ae4ff0f990a"
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint("clear_cuts_user_id_fkey", "clear_cuts", type_="foreignkey")
-    op.drop_column("clear_cuts", "user_id")
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column(
-        "clear_cuts", sa.Column("user_id", sa.INTEGER(), autoincrement=False, nullable=True)
-    )
-    op.create_foreign_key("clear_cuts_user_id_fkey", "clear_cuts", "users", ["user_id"], ["id"])
-
-    # ### end Alembic commands ###
